refactor(js): move around stimulus initializers
